[feat] 사용자 자동 인증 및 헤더 db user_id 자동 삽입 부분 추가 #37
1. 자세한 사용방법 lib/minie/AuthAPI 참조 바람

+++ app/api/cart/delete_all/route.ts
@@ -1,14 +1,17 @@
 import { supabase } from "@/lib/supabase"
-import { NextResponse } from "next/server"
+import { NextRequest, NextResponse } from "next/server"
 
-export async function DELETE() {
-     const tempUid = 1
-    /* TODO: getServerSession(authOption)와 같이 로그인 로직 완성시 얻는 uid 불러오기 로직 */
+export async function DELETE(request: NextRequest) {
+     const uid = request.headers.get('X-User-ID');
+
+    if(uid === null || uid === ""){
+        throw Error("[server]로그인 된 user 정보가 없습니다.")
+    }
 
     const {error} = await supabase
     .from('carts')
     .delete()
-    .eq('user_id', tempUid)
+    .eq('user_id', Number(uid))
     
     if(error){
         return NextResponse.json({ error: 'Failed to delete all cart item: ' + error.message }, { status: 500 })
